Mapeo Answer

+++ src/main/java/com/bgg/quizback/controller/QuestionController.java
@@ -1,5 +1,6 @@
 package com.bgg.quizback.controller;
 
+import java.util.Optional;
 import java.util.Set;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,9 +12,12 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.bgg.quizback.component.mapper.question.QuestionMapper;
+import com.bgg.quizback.component.mapper.question.QuestionAnswerMap;
+import com.bgg.quizback.component.mapper.question.QuestionMap;
 import com.bgg.quizback.dao.QuestionDAO;
+import com.bgg.quizback.dto.QuestionAnswerDTO;
 import com.bgg.quizback.dto.QuestionDTO;
+import com.bgg.quizback.exception.QuestionNotFoundException;
 import com.bgg.quizback.model.Question;
 import com.bgg.quizback.service.QuestionService;
 
@@ -27,7 +31,10 @@ public class QuestionController {
 	QuestionDAO questiondao;
 	
 	@Autowired
-	QuestionMapper questionmapper;
+	QuestionMap questionmapper;
+	
+	@Autowired
+	QuestionAnswerMap questionanswermap;
 	
 	@RequestMapping(value = "/question", method = RequestMethod.GET)
 	public Set<QuestionDTO> findAll(@RequestParam(defaultValue = "0", required = false) Integer page,
@@ -39,29 +46,44 @@ public class QuestionController {
 	
 	@RequestMapping(value = "/question", method = RequestMethod.POST)
 	public QuestionDTO create(@RequestBody QuestionDTO dto){
-		final Question question = questionmapper.mapper(dto);
+		final Question question = questionmapper.dtoToModel(dto);
 		final Question createquestion = questionservice.create(question);
-		return questionmapper.mapper(createquestion);
+		return questionmapper.modelToDto(createquestion);
 	}
-	
+
 	@RequestMapping(value = "/question/{idQuestion}", method = RequestMethod.DELETE)
-	public void deleteQuestion(@PathVariable("idQuestion") Integer idQuestion) {
+	public void deleteQuestion(@PathVariable("idQuestion") Integer idQuestion) throws QuestionNotFoundException{
+		Optional<Question> q = questionservice.findById(idQuestion);
+		if(!q.isPresent())
+			throw new QuestionNotFoundException();
+
 		questionservice.delete(idQuestion);
 	}
 	
+	@RequestMapping(value = "/quiz/{idQuiz}/question", method = RequestMethod.GET)
+	public Set<QuestionAnswerDTO> findByidUserCourse(@PathVariable Integer idQuiz) {
+		 final Set<Question> q = questionservice.findByIdQuestionQuiz(idQuiz);
+		return questionanswermap.modelToDto(q);
+	}
+	
+	@RequestMapping(value = "/quiz/{idQuiz}/question/{idQuestion}", method = RequestMethod.GET)
+	public QuestionAnswerDTO findByidUserCourse(@PathVariable Integer idQuiz, @PathVariable Integer idQuestion) {
+		 final Question q = questionservice.findByIdQuestion(idQuiz,idQuestion);
+		return questionanswermap.modelToDto(q);
+	}
+	
+	
 	
 	@RequestMapping(value = "/question/{idQuestion}", method = RequestMethod.PUT)
-	public void updateQuestion(@PathVariable Integer idQuestion, @RequestBody QuestionDTO dto) {
+	public void updateQuestion(@PathVariable Integer idQuestion, @RequestBody QuestionDTO dto) throws QuestionNotFoundException{
+		Optional<Question> q = questionservice.findById(idQuestion);
+		if(!q.isPresent())
+			throw new QuestionNotFoundException();
+		
 		Question question = questionservice.findByIdQuestion(idQuestion);
 		question.setStatement(dto.getStatement());
 
 		questionservice.update(question);
 	}
 	
-	@RequestMapping(value = "/quiz/{idQuiz}/question/{idQuestion}", method = RequestMethod.POST)
-	public void saveQuizQuestion(@PathVariable Integer idQuestion, @PathVariable Integer idQuiz){
-		questionservice.saveQuizQuestion(idQuestion,idQuiz);
-	}
-	
-
 }
